// src/app/components/value-added/models/value-added.model.ts
export class ValueAdded {
  id: any;
  service_name_en: string;
  service_name_ar: string;
  fee: number;
  img_url: string;
  working_days: string;
  working_timeslot: string;
  decription: string;
  term_conditions: string;
  date_created: any;
}

export class ValueAddedSchedule {
  timeslot_working: string;
  working_days: string;
}